requirements.txt
streamlit==1.28.1
transformers==4.35.2
torch==2.1.1
requests==2.31.0
sqlalchemy==2.0.23
streamlit-authenticator==0.2.3
PyYAML==6.0.1
python-dotenv==1.0.0
nltk==3.8.1
numpy==1.24.3
pandas==2.1.3
plotly==5.17.0
Pillow==10.1.0
